more refactoring for admin view stuff
